requirements.txt
apispec==6.8.2
async-timeout==5.0.1
backports-datetime-fromisoformat==2.0.3
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
flask-smorest==0.46.1
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
redis==6.1.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
webargs==8.7.0
Werkzeug==3.1.3
zipp==3.21.0
